requirements.txt
fastapi
uvicorn[standard]
peewee
pyzmq
pymupdf
scikit-learn
python-multipart
itsdangerous
passlib[bcrypt]
httpx